requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.1
attrs==25.4.0
backtesting==0.6.5
beautifulsoup4==4.14.2
black==25.11.0
bokeh==3.8.1
bottle==0.13.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
contourpy==1.3.3
cramjam==2.11.0
curl_cffi==0.13.0
cycler==0.12.1
dateparser==1.2.2
decorator==5.2.1
dill==0.4.0
executing==2.2.1
fastapi==0.122.0
fastparquet==2024.11.0
fonttools==4.60.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.10.0
greenlet==3.2.4
h11==0.16.0
idna==3.11
ImageIO==2.37.2
ipython==9.8.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
lightweight-charts==2.1
llvmlite==0.44.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
multidict==6.7.0
multitasking==0.0.12
mypy_extensions==1.1.0
narwhals==2.12.0
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.3.3
pandas-ta==0.4.71b0
parso==0.8.5
pathspec==0.12.1
peewee==3.18.3
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
plotly==6.5.0
polygon-api-client==1.16.3
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.1
proxy_tools==0.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
PyQt6==6.10.1
PyQt6-Qt6==6.10.1
PyQt6_sip==13.10.3
PySide6==6.10.1
PySide6_Addons==6.10.1
PySide6_Essentials==6.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.3.0
pytz==2025.2
pywebview==6.1
PyYAML==6.0.3
QtPy==2.4.3
regex==2025.11.3
requests==2.32.5
schedule==1.2.2
scikit-learn==1.8.0
scipy==1.16.3
shiboken6==6.10.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==0.50.0
threadpoolctl==3.6.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.38.0
vectorbt==0.28.2
wcwidth==0.2.14
websockets==15.0.1
widgetsnbextension==4.0.15
xyzservices==2025.11.0
yarl==1.22.0
yfinance==0.2.66
